Move filter_data_categories back to "graph_task.py" so the diff is easier to review - move consolidate_query_matches into its own module.
