+++ src/fidesops/service/privacy_request/request_runner_service.py
@@ -20,10 +20,10 @@ from fidesops.models.policy import (
 )
 from fidesops.models.privacy_request import PrivacyRequest, PrivacyRequestStatus
 from fidesops.service.storage.storage_uploader_service import upload
-from fidesops.task.filter_results import filter_data_categories
 from fidesops.task.graph_task import (
     run_access_request,
     run_erasure,
+    filter_data_categories,
 )
 from fidesops.tasks.scheduled.scheduler import scheduler
 from fidesops.util.async_util import run_async
