+++ tests/integration_tests/test_mongo_task.py
@@ -13,8 +13,6 @@ from fidesops.graph.config import (
     ScalarField,
     Collection,
     Dataset,
-    CollectionAddress,
-    FieldPath,
 )
 from fidesops.graph.data_type import (
     IntTypeConverter,
@@ -32,9 +30,7 @@ from fidesops.models.privacy_request import PrivacyRequest
 from fidesops.schemas.dataset import FidesopsDataset
 from fidesops.service.connectors import get_connector
 from fidesops.task import graph_task
-from fidesops.task.filter_results import (
-    filter_data_categories,
-)
+from fidesops.task.graph_task import filter_data_categories
 
 from ..graph.graph_test_util import assert_rows_match, erasure_policy, field
 from ..task.traversal_data import (
