+++ tests/task/test_filter_results.py
@@ -7,8 +7,8 @@ from fidesops.graph.config import FieldPath, CollectionAddress
 from fidesops.task.filter_results import (
     select_and_save_field,
     remove_empty_objects,
-    filter_data_categories,
 )
+from fidesops.task.graph_task import filter_data_categories
 
 
 def test_select_and_save_field():
